chore(folder-structure): Move hooks and hocs to respective apps
move hocs fix
