+++ src/app/content/App/Account/About/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { compose } from 'redux';
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import { ExternalLink, Time } from 'components/atoms';
 import { WEBSITE_DOMAIN } from 'app/lmem';
 import Title from './Title';
