+++ src/app/content/App/Account/Filters/Filters.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import ContentTitleTop from 'components/organisms/Notification/ContentTitleTop';
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 
 export const Filters = () => {
   const { t } = useTranslation();
