+++ src/app/content/App/Account/Menu/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Link, RouteComponentProps } from 'react-router-dom';
 import { ExternalLink, OpenButton } from 'components/atoms';
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import { WEBSITE_DOMAIN } from 'app/lmem';
 import Account from './Account';
