+++ src/app/content/App/Contribute/ContributeScreen/index.ts
@@ -1,4 +1,4 @@
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import ContributeScreen from './ContributeScreen';
 
 export default withTitle('title.post_form')(ContributeScreen);
