+++ src/app/content/App/Contribute/PreviewScreen/withTitle.ts
@@ -1,4 +1,4 @@
-import withTitle from '../../../../hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import { PreviewScreenOwnProps } from './PreviewScreen';
 
 export default withTitle<PreviewScreenOwnProps>('title.preview');
