+++ src/app/content/App/Help/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { List, ExternalLink, Anchor } from 'components/atoms';
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import { WEBSITE_DOMAIN } from 'app/lmem';
 import Content from '../Account/About/Content';
 import Container from './Container';
