+++ src/app/content/App/Question/FormScreen/index.ts
@@ -1,4 +1,4 @@
-import withTitle from 'app/hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import FormScreen from './FormScreen';
 
 export default withTitle('title.ask_info_form')(FormScreen);
