+++ src/app/content/App/Subscriptions/index.tsx
@@ -1,5 +1,5 @@
 import { compose } from 'redux';
-import withTitle from '../../../hocs/withTitle';
+import withTitle from 'app/content/hocs/withTitle';
 import withConnect from './withConnect';
 import Subscriptions, { SubscriptionsScreenProps } from './Subscriptions';
